refactor: 최종 결과 화면 컴포넌트 분리
