+++ apps/web/src/components/midpoint-result/molecules/MapHeader.tsx
