+++ apps/web/src/components/midpoint-result/organisms/styles.css.ts
@@ -2,7 +2,6 @@ import { theme } from "@repo/ui/tokens";
 import { zIndex } from "@repo/z-index";
 import { style } from "@vanilla-extract/css";
 import { recipe } from "@vanilla-extract/recipes";
-import { keyframes } from "@vanilla-extract/css";
 
 export const containerRecipe = recipe({
   base: {
@@ -30,28 +29,6 @@ export const wrapper = style({
   width: "100%",
 });
 
-export const container = style({
-  display: "flex",
-  justifyContent: "center",
-  alignItems: "center",
-  height: "36px",
-  borderRadius: "100px",
-  padding: "12px",
-  cursor: "pointer",
-  backgroundColor: theme.colors.gray0,
-  whiteSpace: "nowrap",
-});
-
-export const locationTextStyle = style({
-  margin: "0 8px 0 10px",
-  color: theme.colors.gray90,
-});
-
-export const stationTextStyle = style({
-  marginLeft: "8px",
-  color: theme.colors.gray95,
-});
-
 export const bannerContainerStyle = style({
   position: "relative",
   width: "100%",
@@ -72,30 +49,6 @@ export const deleteBtnStyle = style({
   top: 12,
   cursor: "pointer",
   zIndex: zIndex.overlay,
-});
-
-export const headerStyle = style({
-  backgroundColor: theme.colors.gray0,
-  position: "fixed",
-  top: 0,
-  width: "100%",
-  maxWidth: "600px",
-  height: "58px",
-  zIndex: zIndex.floating,
-  display: "flex",
-  alignItems: "center",
-  justifyContent: "center",
-  cursor: "pointer",
-});
-
-export const backBtnStyle = style({
-  position: "absolute",
-  left: "8px",
-});
-
-export const titleStyle = style({
-  flex: 1,
-  textAlign: "center",
 });
 
 // TODO: 결과 화면 구현중
@@ -160,13 +113,6 @@ export const patternWrapperStyle = style({
   height: "100%",
 });
 
-export const transportContainerStyle = style({
-  display: "flex",
-  alignItems: "center",
-  justifyContent: "center",
-  zIndex: zIndex.toast,
-});
-
 export const lineIndicatorStyle = style({
   display: "flex",
 });
@@ -184,51 +130,8 @@ export const lineNumberStyle = style({
   fontWeight: "bold",
 });
 
-export const progressBarStyle = style({
-  height: "14px",
-  flexGrow: 1,
-  borderTopRightRadius: "10px",
-  borderBottomRightRadius: "10px",
-  position: "relative",
-  display: "flex",
-  alignItems: "center",
-  justifyContent: "center",
-  zIndex: 1,
-  transform: "translate(-2px, 0)",
-});
-
 export const timeDisplayStyle = style({
   fontSize: "20px",
   fontWeight: "bold",
   flexShrink: 0,
-});
-
-export const AddLocationButtonStyle = style({
-  background: theme.colors.orange40,
-  padding: "0 12px",
-  width: "150px",
-  height: "36px",
-  borderRadius: "100px",
-  justifyContent: "center",
-  alignItems: "center",
-  cursor: "pointer",
-});
-
-const shimmer = keyframes({
-  "0%": {
-    backgroundPosition: "200% 0",
-  },
-  "100%": {
-    backgroundPosition: "-200% 0",
-  },
-});
-
-export const skeletonLoading = style({
-  width: "45px",
-  height: "16px",
-  marginLeft: "8px",
-  borderRadius: "4px",
-  background: `linear-gradient(90deg, ${theme.colors.gray20} 25%, ${theme.colors.gray30} 50%, ${theme.colors.gray20} 75%)`,
-  backgroundSize: "200% 100%",
-  animation: `${shimmer} 1.5s infinite`,
 });
