+++ apps/web/src/components/midpoint-result/templates/style.css.ts
